Issue 13: refactor: Move Query mock support classes to sub package "query" as in javax.jcr
