+++ magkit-test-jcr/src/main/java/de/ibmix/magkit/test/jcr/query/QueryManagerStubbingOperation.java
@@ -1,4 +1,4 @@
-package de.ibmix.magkit.test.jcr;
+package de.ibmix.magkit.test.jcr.query;
 
 /*-
  * #%L
